From svn r35952 on 2008-06-26 11:10:29 -0500 (Thu, 26 Jun 2008)
It will now be more efficient when retrieving objects where the get() is on
a more generic class, but the objects it's getting back need to be subclassed.
When it detects an item that belongs in a subclass, it will start a new
DB statement handle executing SQL that contains the original query, plus
whatever joins are required to also pull in the properties defined in the
subclass/table, and reads data from it in parallel with the original statement
handle.

That means that db-level iterators now return arrayrefs of data that can
be sliced to pull out data from the columns that go with the respective class'
properties that make up the object, instead of the old way that just returns
a hashref that can be easily blessed into the right class.

RemoteCache datasources are broken by this change and will be fixed soon.