From svn r36349 on 2008-07-09 10:28:24 -0500 (Wed, 09 Jul 2008)
Removed a shortcut during checking for previously done queries.  If the
number of params in the request was less than two, it assummed the lone
param left was the ID param, and another part of the caching system was
picking it up.

This assumption wasn't true for the case where a query was done on a
class with one param that wasn't the ID param, and then later a get() on
a subclass with the same param comes along.