From svn r38466 on 2008-09-09 16:33:41 -0500 (Tue, 09 Sep 2008)
Fix for cross-datasource joins when the secondary data source is
a factory class:
We can properly create the secondary loading template by putting
if its contstriction until later.  At that point, the secondary
rule is available, and we can let the factory class properly
resolve to a concrete class before creating the loading template

When loading object from the DB iterators, get the ID properties
using a hash slice of the proto-object hash instead of the DB row listref.
That way any constant property values are available