From svn r38659 on 2008-09-16 17:05:58 -0500 (Tue, 16 Sep 2008)
Updates to the handling of the has-many relationship.  Now takes an arrayref during construction.  Allows members of the arrayref to be full object descriptions, or just the part of the id which cannot be inferred.  Now supports read-write accessors on indirect properties which do "the right thing", adding/removing bridges as needed.