From svn r38802 on 2008-09-18 23:23:21 -0500 (Thu, 18 Sep 2008)
BUG FIX: The latest update to handle indirect accessors broke some old features, while adding new ones.  We now handle indirect properties which are read-write regardless of whether the target value is part of the id of the target.  Test case for both added.

NEW FEATURE: A has-many relationship with primitves, non-UR objects, or UR objects w/o foreign-key linkage now works, doing what you'd expect internally: storing an arrayref in the hash for the object.  There is a bug that these are not handled in in-memory transactions yet.