From svn r43003 on 2009-01-26 15:18:52 -0600 (Mon, 26 Jan 2009)
Before, the class initializer was faking a table_name property
in some cases, such as for inline-defined data sources, so that
the class' properties would get column_names filled in.

This caused problems with classes that don't have their own
table_name, but inherit it from their parent.

The new way is to have the property normalizer fill in
column_name if its datasource isa UR::DataSource::File