From svn r44621 on 2009-03-12 16:17:57 -0500 (Thu, 12 Mar 2009)
now splits a line into and sets array length equal to the # of expected columns.  This will ensure that extra columns in the datasource are exposed in the last column property, and that constant_value values will not end up in another property due to truncated trailing null columns, or columns missing altogether.  Setting the LIMIT arg on the split should marginally speed up this operation