From svn r44997 on 2009-03-24 17:19:40 -0500 (Tue, 24 Mar 2009)
Fixes a bug where if you do a query, but ask for an iterator instead
of getting all objects as a list, and read some (but not all) the objects
from the iterator.  Later, do a get() (or iterator) with the same params.

The old behavior would, for the second get(), only return whatever
objects were fetched from the first iterator.

The fix involves the object fabricator closure keeping a record of
planned changes to $UR::Context::all_params_loaded as objects are passed
up the chain.  When the underlying context closure (which calls the object
fabricator closure) reaches the end of the DB-loader iterator, apply the
all_params_loaded changes by calling commit() on the fabricator closure.