From svn r45356 on 2009-04-03 08:43:33 -0500 (Fri, 03 Apr 2009)
Added the optional "valid_values" parameter to class property definitions to support simple enumerated types.  Renamed "sub_classification_property_name" to "subclassify_by" to fit with the style of the rest of the class definition.  Updated the logic to support the simplest possible case of the property right on the object returning the explicit class name.  The other two sub-classification properties could be effectively re-framed under this and eliminated from the class definition.