From svn r46194 on 2009-04-23 18:29:31 -0500 (Thu, 23 Apr 2009)
Fixed a bug where getting MetaDB objects via their 'namespace' property would
fail.

The cause was that 'namespace' was a 'via' property through datasource_obj,
and the Context thought that meant it was a cross-datasource join.

The fix was to change 'namespace' to be a calculated property based on the
data_source