From svn r47134 on 2009-05-21 21:40:04 -0500 (Thu, 21 May 2009)
Major pruning of the parts of UR::Object we do not want to have to support after public release.  Created UR::ObjectDeprecated for the methods UR still uses but which we want to eliminate.  Created UR::ObjectV001 for the parts of the API which UR no longer uses, but other namespaces (Genome) do use still.  Those namespaces use that module explicitly to get backward compatibility.