From svn r48934 on 2009-07-15 11:43:31 -0500 (Wed, 15 Jul 2009)
Fixed a bug that happens when doing a get() on an abstract base
class, using only 'id' as a filter.  When objects are loaded from
the SQL, they need to get subclassed via a subordinate iterator, but
the rule for that was not normalized, and so only 'id' remained as
the filter, not the subclass_id property as well.

UR::DataSource::RDBMS::_generate_template_data_for_loading() assummes that
if a class has more than one ID property, then we can remove the filter
called 'id' because the subclass_id will also appear in the rule.  The
end result was that the query for subclasses data had no filters at all.

The data returned to the user got filtered out downstream, but it made
the process less efficient