From svn r49426 on 2009-07-30 16:30:54 -0500 (Thu, 30 Jul 2009)
Fixed a bug where the order of items returned from a get() with -order
could be different whether the items are retrieved from the DB or the
object cache.  The sorter for object cache items included 'id' as the
last sort property, but the generated SQL did not include id columns.

The fix involved having the class data and loading templates keep track
of the order by columns as a listref instead of the order by clause as
a string.  Then, the user's order by list can be prepended to the front
of the list when necessary.  Finally, the SQL order by clause is constructed
at the last moment.