From svn r49577 on 2009-08-05 11:22:44 -0500 (Wed, 05 Aug 2009)
Fixes a bug where an iterator's next() method could skip relevant objects
or include wrong objects.  This can happen if the objects' states change
between the time the iterator is created and when next() is called.

This change means that an iterator will return objects matching the rule
when the iterator is created.  It also means that next() can return an
object that does not match the rule if it would have matched at the iterator's
creation time.  Finally, the loading iterator (called by next()) can throw an
exception if it is about to return a deleted object - meaning that it is
an object that matched the rule at the iterator's creation time, but
has been deleted in the meantime.

The mechanism is that the underlying context iterator will not return objects
that already exist in the current Context's object cache.  The Context
maintains a list of open loading iterators.  When the loading iterator gets
an object from the underlying context iterator, it sees if that new object matches
the rules for the other loading iterators, and inserts it into their cached
objects list if so.  Finally, the loading iterator is blessed so that a destructor
gets called at the appropriate time and cleans itself out of the open iterator
list.