From svn r50303 on 2009-08-25 10:41:15 -0500 (Tue, 25 Aug 2009)
Fix a bug in the loading iterator.

Loading an object with hints can cause the same original object to be loaded
while satisfying the hints.  By the time the topmost-level
underlying_context_iterator tries to return the original object, it's
considered already loaded, and so returns nothing.  And because we're already
tried to pull an object from the @$cached list and found nothing, the iterator
bails out.

By reading from the underlying_context_iterator first, another loading iterator
can inject this original object into our @$cached list and we'll see it even
though the underlying_context_iterator still returns undef (again, because it's
already loaded)