From svn r50635 on 2009-09-01 16:55:54 -0500 (Tue, 01 Sep 2009)
Fix for writing/merging file datasources in _sync_database()

When using quick_rename, the temp file is now made with IO::File->new
instead of File::Temp.  That should mean that when the temp file
is renamed over the existing file, it should have permission appropriate
for the user's umask.