Don't install the singular accessor sub when the property name is not plural
For is_many properties where the singular_name and plural_name are the
same, which happens when the property name is not a plural word, skip
inserting a singular accessor sub.  Previously, this singular sub
got the same name as the plural sub, and so replaced it.  The change
means the plural sub is now installed, which is more likely to be
what the user wanted