Fix a bug involving filters on delegated properties with overlapping classes/tables.
Where a class A has a delegated property to class B and another delegated
property to class C (through B), and then you do a get() on A filtered by
both of these properties.  Previously, it would come up with SQL joining to
the wrong table for the second property.