Fix for my previous fix/commit "Better error msgs"
The previous solution was to immediately throw an exception if there
was a problem use-ing $r_class_name or getting class metadata for
$r_class_name.  Turns out that's bad because the primitive type
names like 'String' and 'Integer' don't have packages or metadata.

The fix is to remember any exceptions thrown while use-ing or getting
metadata.  Then, if $reverse_as was supplied or derived, and there's
no r_class metadata, go ahead and throw an excption with the saved error
from before.