Change the rules for data-change conflicts when loading data
Before, when data was loaded from the DB matching an object already
in the cache, an exception was raised if the object's property had been
changed and DB data had changed since the object was loaded.

Now, it also allows the change if the newly loaded data matches the
current value of the property, or if it matches the db_saved_uncommitted
value.

It also makes the exception message clearer by explicitly saying if
any of the values is undef, where before that sould show up as the
empty string.