Change for SQL where clause building where property exists in parent and child classes
If a filtered property exists in both a parent and child class, then the where clause
will have a chance to be applied to all tables in the hierarchy where that proeprty
has a column.

Previously, it would only apply where where clause to the first class with a matching
property, even if it didn't have a column at that level in the class hierarchy.