Change accounting of applicable filters during SQL generation
Instead of marking a flag every time a filter matches (%satisfied_filters)
and then comparing the set flags to the initial filter list (%filters)...

Make a copy of the initial filters into %filters_to_satisfy, and delete
keys as the filters are applied.  When we're done, if there is anything
left in %filters_to_satisfy then there's an error.

Should provide a minor speed boost