Improve SQL performance when querying a subclass and the parent class has subclassify_by
A query on a subclass where the parent class is_abstract and has a subclassify_by property
(meaning that the parent class has a property which directly stores the proper subclass for
each row - subclasses inherit the property from the parent, and the subclass isn't is_abstract)
should have a filter added to the rule to keep only rows of the subclass we're interested in.
This will improve the SQL performance when it's later constructed.