Rearrange autogenerated help text for object lister commands
Moved the text for filterable properties to the bottom, since
that's most relevent to the user

Use help_synopsis if it exists for the examples section, otherwise
use some generic text showing multiple filters and/or shows

In the list of filterable properties, it includes inherited properties.
Does not include private (starts with '_') property names, properties
whose data_type includes '::' (object types), and filters out duplicate
names.

If a property is undocumented in a child class, but is in a parent, use
the parent's documentation in the help text.  This applies both to the
properties of the command module, and to the properties of the subject_class_name
of an object lister command