Attempts at getting the View system cleaned up and ready to merge
The ./UR/Object/View/Default/Text.t test is failing because of a recursion
problem where object A refers to a B, and B somehow refers back to an A.
I've added in some attempts at detecting this recursion, but it is still
not optimal.

Maybe it needs a way for aspects to register with the top-level view that
it's rendering some object, and refuse to recurse down into it if that
object is in the process of being rendered by an encompassing view.