Fix for spurious exception about: Can't locate object method "class"
The Genome-namespace test Genome/Model/Build/Command/Status.t was dying with
the exception
    Can't locate object method "class" via package "Genome::Model::Command::Build::VerifySuccessfulCompletion"

I'm not sure what changed to trip this bug, but it's related to commit 3ea4c43c in
_first_class_in_inheritance_with_a_table().  An "eval qq()" was commented out, but
not the die immediately after, which was originally there to catch a syntax error
in the eval-ed code.  During the above test, $@ was still set to something by
a previous eval (probably something in the module loader).

The fix is to just comment out the die here.