Attempt to mitigate slowness in UR::BoolExpr by making query() smarter
In a basic id-only get()/query(), the previous behavior was to look
directly into the object cache for the queried class and ID, and return a
matching object quickly, before constructing a BoolExpr and performing
a more thorough search in get_objects_for_class_and_rule().

The new behavior also uses searches the object cache for any direct
child classes of the queried class, but not recursively.  The extra time
spent searching additional classes should be worth it compared to the
slowness of UR::BoolExpr::resolve().