Xsl now allows both a perspective and a desired_perspective.
When a controller is trying to create a view for a perspective that doesnt exist and it catches an exception, it may try again by setting the perspective to default and the desired_perspective to what they originally wanted.

When the Xsl view tries to create the Xml view, it will attempt the desired_perspective first, then fall back to its perspective. This allows you to implement an alternate perspective with only the xsl template.  It will fall back to Default/Xsl.pm to do the transform, while still using an Xml view that may be available for the perspective.