Hack to get around a problem with OSX 10.5 and getline()/<FH>.
Perl 5.8.8 that comes with OSX 10.5 has peculiar behavior with getline/<FH>.
Other platforms (many Perls on Linux, Perl 5.10 on OSX 10.6, etc) set $! to 0
after a getline returns undef when you're at a file's EOF.  That way you can
distinguish between EOF and a real error condition.

The problematic Perl does not set $! to 0, and so at EOF, $! retains whatever
value it had after the last failed syscall(), which is usually a failed stat()
while searching @INC from loading a module.

The hack is to set $! to 0 by hand before the getline.  If the getline
triggers an underlying sysread() to fill the buffer, and there's an error,
then $! should get set to something before returning.