Fix a bug with loading and saving objects where the table_name included the schema/owner
The SQL that queries the data dictionary for classes that have no
MetaDB previously saved do not work properly if you use the data source's
default schema/owner in the 'owner' field, and then use a table_name that
includes an alternate schema.  For example, the DS's owner/schema is Foo and the
table_name is Bar.TableName.  It now splits the proper schema out from the
compound table name.

A similar problem existed when saving the objects.  It used to construct a table
name like this: Foo.Bar.TableName.  Now, it does the right thing (Bar.TableName).