Support properties that are calculated and not is_mutable.
Properties that are ! is_mutable means that a value for that property must
exist at object creation time.  If there is also a calculation supplied for
that property, then UR::Context::create_entity() will call that calculation
if it wasn't supplied in the creation params.  This sub is called as a
class method, with the class name is $_[0] and the currently-known creation
params as $_[1].

The accessor writer creates these as read-only properties.

Added to the calculate properties test to check for these new immutable+calculated
properties, that they can also have column_names and save to the DB, and that
is_constant+calculated properties act similarly, but the calculation is
called more lazily.