Don't show duplicated params in Command help text
If a parent Command class defines a param as optional/required, and
a child class overrides it to be the opposite, the help text would
list that param twice, once in Required Params and once in Optional.

This fixes the problem by only considering the property defined on the
most-specific class.