Better error message while processing an indirect property
When changing an indirect property, and that change necessitates creating
a new object to hold the change, and the 'to' meta-property is incorrect,
it would previously print out an ugly stack dump that didn't describe the
problem at all.  The new code wraps an eval {} around the adder method
and if there was a problem, it detects this condition and prints a more
useful error.

For example, a class definition like this:

class Foo {
    has => [
        bob => { is => 'Bob', id_by => 'bob_id' },
        bob_color => { via => 'Bob', to => 'color' },
    ]
};
class Bob {
    has => 'size',
};

If you do something like
$foo->bob_color('red');
it would fail with a stack dump complaining about extra params 'color' blah blah

Now it correctly sees that the Bob class has no property called 'color'.
This is more likely to occur when the user ommits the 'to' meta-property, and the
system auto-assigns 'to' to be the same name as the originating property.