A calculated subclassify_by now works for objects loaded from the DB.
Changed the calling conventions for these calculated subclassify_by at
create() time.  The old behavior was to call the function as a class
method, with the class_name as the first arg, and the other args are
whatever was passed in to create().

The new behavior is to fake-up an object blessed into the parent/abstract
class with data from the create() args, then call the subclassify_by
method as an object method.  This will parallel the behavior in the
Context's object fabricator, and let the user put in metadata like
calculate_from