Updated the indirect accessor test for things which already worked but were not tested properly
This test had a SKIP: section and erroneously claimed a create() like this
always created a new foreign object.  In actuality, it creates one if the
thing you're trying to point to does not exist, and links up to an existing
one if it does.

The new tests try both cases