More changes to subclassify_by final subclass resolution
Went back to using _create_entity_from_abstract_class() near the top of
create_entity(), and it's able to handle cases where you pass in a
value for the subclassify_by property in the creation rule, even if
it's an indirect property (though linking to the other object isn't
quite working).

Calculated subclassify_by with a subref works, when you specify all the
params in that property's calculate_with in the creation params.
subclassify_by with a default_value is also working.