Fix for indirect property value handling in create_entity
It was incorrectly traversing the complete path to the final property
location and then trying to set the local linking properties to this
final indirect property.  That's the wrong thing to to in the case
of a multiply-indirect property.  Go back to only traversing one level
of indirectness, and the (possibly recursive) call to $foreign_property->create()
will handle multiple levels of indirectness behind the scenes.