Revert an earlier change when we needed to keep around %extra params after rule creation on an abstract parent class
An earlier version of the calculated_subclassify_by branch would create an
instance of the abstract parent class before determining the proper subclass,
and so needed to remember %extra params that might only apply to subclasses.

We're not doing that anymore, and capturing %extra was coving up cases where
someone might pass in params that don't apply to a class.