UR::DataSource::SQLite::_init_database() is now called from within create_dbh() instead of server()
It's more logically consistent to put off initializing the DB until we actaully
need the handle.

For the common case where a SQLite-based datasource overrides server() to
configure the DB file, this has the side effect of requiring the file to
actually exist before we can get the handle, since _init_database() throws
an exception of the DB file, dump or schema does not exist.

Some testcases altered to touch an empty DB file to fill the requirement