First pass at the new caching/seeking implementation
The new behavior works by keeping a cache of post-split file data (as an
arrayref) and the file seek position that data can be read from (ie. the
tell() from before we did a getline()).

Before entering the loading iterator, it searches the cache for the greatest
file offset where the sorted columns data is less than what appears in the
rule.  If no matches are found, it starts at position 0.

Each loading iterator is allocated its own slot in the cache to perform
updates on.

Each time the iterator is entered, it has a chance to seek() using the
last saved position if other iterators are working on the same file.

After finding a matching object, the loading iterator checks to see if
the data in the sorted columns is different than what it has stored in
its cache slot.  If so, it updates the cache-saved position with the
saved file position from the last trip through the iterator, which
ends up being the file position from just before the last read.