Fix an error in SQL generation involving indirect properties with tree-like (rather than linear) inheritance
The git() in question involves a filter on a doubly indirect property, where
each class in the indirection also has inheritance.  Something like this
where '|' is the indirection and '=' is inheritance

my @matching = FilterClass->get(propertyA => 'Some value');

  Filter Class
  property A
     |
     |
  related class  ===   Related class parent
  property A
     |
     |
   related class 2 ===  Related class 2 parent
   property A
  final location

The old code assummed all joins involved in resolving an indirect property
were meant to be chained together.  It fails in this case because as it adds
the join to RelatedClassParent because of inheritance on RelatedClass, but
when it adds the join to RelatedClass2, the join is incorrectly written
to be to RelatedClassParent instead of RelatedClass.

This new code keeps track of two join "pointers", one because of indirection
and another from inheritance.

A UR test is still failing (49_complicated_get.t), so this isn't a complete fix yet