Fix a bug where the file offset cache could be used incorrectly
Require the comparison show the candidate offset is strictly before what
we're looking for, and not just that the comparison returns true (0).  It
would incorrectly choose to take a cache hit when the comparison was less-than
and the comparator returned 0, indicating that candidate was indeed less than,
but it didn't guarrantee that position was before anything that could match.