Fix a bug in UR::Context object cache pruner
If the memory pressure is so great that the pruner is getting called for
every get(), the incrementer value in the pruner can be 0.  If the user
process is also holding outside references to objects such that it can't
get blow the lowwater mark, then the pruner can't release objects.  In
this case, the pruner never leaves the main loop and the process becomes
hung