New test for caching of queries with subclass specific params
Subclasses that do not have a dedicated table (e.g., those using
hangoff tables or calculated properties exclusively) are currently
causing problems with query caching. Querying by subclass specific
parameters causes UR to think that it has loaded all objects of
that specific subclass (more preciesly, it thinks it ran the query
with none of the subclass specific parameters). This test exposes
the problem so a fix can be developed.