Fix a problem with the Tk debugger running away
A construct like
eval "use NonExistantModule";
will cause the Tk debugger (Devel::ptkdb) to miss stopping at the next
statement if you do a step-over, and something in the chain of function
calls does one of those string evals.  It runs away because of the way
the debugger tracks function call depth.  The function entry is recorded,
but the exit is not tracked because the exception returns up more than
one call frame without the debugger noticing.

The problem is really with the debugger not catching exceptions thrown
during a function call in DB::sub(), but my attempts to fix it fall short.

This fix to the class initializer does away with the problem eval, and
actually fixes a hole in the logic around making sure modules are
loaded from a subdir of where the Namespace module was loaded from.
Previously, is a class specified a parent class, and that parent class
did not exist under the Namespace dir, but did exist somewhere else in
@INC, it would go ahead and load it anyway.