Defer loading subcommands until they are needed.
Instead of loading all dynamic subcommands at compile time for the
parent command, use __extend_namespace__ to allow them to be created
at the time first referenced. Still compile all of them in situations
where the complete list of subcommands is required, but defer that
work until the information is first requested.

This speeds up, for example, the generation of help text for a command
that has a subcommand that dynamically generates its subcommands, by
not requiring the loading of a second layer of commands.