Clean up the 98_ur_update.t test
Don't try to make UR.pm writable.  I have no idea why this was in there in
the first place.

Don't try to touch the SomeSQLite DB file.  Code in the datasource does that
automatically now my creating the DB in a temp dir

Don't remove the MetaDB database or dump.  They're also created on the fly
in a temp dir