Improve the indirect read-only accessor in certain cases
The default action is still to retrieve the bridge objects, then
iteratively call the 'to' property on them to get the list of results.
This can be excessivly slow if that 'to' property is also delegated
through another object accessor.  It will make a series of one-off
queries to the DB.

If appropriate, it will now retrieve the bridge objects, then extract
a list of values linking the bridges to the final objects, then do a get()
for final objects using an in-clause with the linking values.  Finally, call
the last 'to' property on those to get the list of results.