Improve speed of _loading_was_done_before_with_a_superset_of_this_rule()
When a template has had many (thousands) of rules applied, this method gets
slow because it has to check against thousands of rules.  Make it faster
by making a regex out of the rule's values and apply the regex against
the value_id's in the all_params_loaded query cache.  Those that pass the
regex get the full parameter checking