Fixed 'ur update classes-from-db' from throwing bad exceptions
While processing unique constraints, it first tries to remove UR
metadata about constraints, then adds them back in.  IF there's a
new constraint, it can't remove metadata that doesn't exist yet,
and it threw an exception.  In this case, it's OK not to die
if it's trying to remove a non-existent constraint