Refactor the underlying_context_iterator to return objects after ancillary data is loaded
Cleaned up the underlying_context_iterator to hold on to an object longer.
If the columns representing the primary asked-for object come through
more than once (which happens if there's an SQL join involved from delegated
or hinted properties), then don't return that object until all the joined-to
data has been loaded.  It does this by holding onto that object until
those columns represent the next object, then returns the object it was
holding on to.

Tests are still failing with this code; it's not ready yet