Accessors detect value changes by wrapping the comparison in an eval {}
Some values can be objects with overload logic that dosen't properly
handle being compared to some other kind of item (otner kind of object,
numbers, undef, etc).  Most accessors use 'ne' to see if the value
has changed to determine whether to change the data in the object and
send a signal_change.

Wrap the comparison in an eval {} and do the setting/signalling if
the values are different or the eval thre an exception about overloading
failure.