For Ghost classes, don't blindly make properties is_optional
Previously, when a Ghost class got defined, it would copy properties from
the live-class to the Ghost class, but make all non-id properties of the
Ghost class is_optional => 1.

This ended up causing problems when deleting rows when a column was NOT NULL
and part of a foreign key, since deletion logic in UR::DataSource::RDBMS would
first try setting the column to NULL (to avoid circular constraints) before
deleting rows.  In the problem case, the property of the Ghost object says it
is is_optional, so the _sync_database logic tries to set it to NULL, which
fails because of the NOT NULL constraint