Fix a problem where the 98_ur_update.t test can fail
If you run the test from somewhere other than under UR's t/ subdir,
the test can fail in one of the places where it tries to retrieve
Table metadata object for the PERSON table.

It seems to be a strange interaction between Ghosts, new objects commit()ted
when UR_DBI_NO_COMMIT is on, and timing.

The second or subsequent time it tries to create/update the PERSON table
and Person class, it can fail to see the table metadata object is changed
because of the way those interactions happen.  When the software transcation
is rolled back, the new PERSON metaobject becomes a Ghost with
db_saved_uncommitted data.  The next time the PERSON metaobject gets
create()d, the Ghost is resurrected and is now a regular, non-new object
with no changes.  And so, the part in UR::NS::Cmd::Update::ClassesFromDb
where it picks out datasource metadata objects with __changes__, it can
fail to see this PERSON metadata.

If works by accident most of the time because the 'last_object_revision'
property gets changed in the course of the test, and then __changes__ finds
it.  But if two parts of the test run through in the same second, then
the object appears unchanged.

We could change the behavior of a transaction undo-ing a new object with
db_saved_uncommitted data, but this issue should only come up in this one
testcase, and the mechanism behind transactions and changed objects is
going to be changing soon, anyway.

The net result is that the update command is getting changed to look
for not only objects with __changes__, but also objects with
db_saved_uncommitted data.