Added method table_and_column_names_are_upper_case() to UR::DataSource
This controls whether the column_name property metadata gets forced to
upper-case at class initialization time.  The default is to make the
column_name the same as the property_name if column_name is not specified.

This will keep out clutter in the class definitions for data sources where
the column_names are all upper-case