Remove wierd uninitialized value warning with DBD::Pg
Private storage data is implemented with Tie magic on db/statement
handles and is probably handled by XS routines.  Setting a private
key to have the value undef generates the warning

Use of uninitialized value in subroutine entry at .../UR/DBI.pm line xxx

http://www.perlmonks.org/?node_id=415183 gives a hind as to what I think
is happening, that the XS code is retrieving the string value of the scalar
value we're trying to set.  That value is undef when called from
UR::DBI::after_all_fetches_with_sth() (when clearing
$sth->fetch_timing_arrayref).  report_uninit() gets triggered by the conversion
of undef to a string, and since the current executing is in XS code, the last
perl opcode executed was enter-sub.

That's a long-winded way of saying that it seems to be a bug in DBD::Pg,
since we don't see this problem with the other DBD drivers.