Easier way to override aitogenerating IDs for a class
You can specify 'id_generator' in the class definition (old name is
id_sequence_generator_name is still accepted as a synonym).

If the class has no data source, and no id_generator, then the default
value is -urinternal, and means to call the method
UR::Object::Type::autogenerate_new_object_id_urinternal().  Also accepted
is -uuid to generate hex UUID strings.

For classes with datasources it acts as it did before, where no value means
look for a sequence with the same name as the table, ending in "_seq".  Or
you can force a specific sequence name.