Detect deleted DB rows and remove cached objects
Object fabricators now collect a list of objects they expect to be loaded
from the data source (and passed to the fabricator to make objects from).
If expected objects are missing, it uses $context->__merge_db_data_with_existing_object
to make the previously-existing object go away.

__merge_db_data_with_existing_object is changed to allow $pending_db_object_data
to be undef, meaning that the DB row is non-existent and the object should
be abandoned