Better detection of changing DB data
The LoadingIterator now does a more thorough job of handling the cases where
database data has changed since we last loaded it.  In particular, it
previously did not handle at all the case where a row was deleted.

In order to detect row deletions, it also had to be changed to handle the
cases where the object still exists in the DB but has been changed to no
longer match the BoolExpr, or if the query specifies order-by and the value
of the order-by columns have changed in either the DB or in memory.

There are still many tests that fail.