Hack to fix _change_count accounting around software transactions
During __merge_db_data_with_existing_object, don't create change records when
the updating the object's current value.  This is to fix an issue where DB data
gets changed, the object is reloaded causing it's value to be changed to match
the database, then a software transaction gets rolled back.  After the rollback,
the object would incorrectly get the value originally stored in the database
because of the change record created inside __merge