Fix "not in" operator
Before, it was returning false if any object value was in the included list

Now, it returns true if any object value does not match any value in the list.
This matches SQLite's 'NOT IN' behavior.

This has been a hidden problem for a while.  It was getting covered up because
prior behavior of the underlying context iterator was to not return objects that
were already loaded