The new deletion detection code changes iterator behavior
Iterators used to collect matching objects when the iterator was created,
so whether or not an object got returned by the iterator was decided when
the iterator was created.

The new deletion detection code means that, in practice, objects get returned
if they match the bx at the time they get pulled off the iterator.  We'll
be more conservative and just say that iterator behavior is undefined when
the set objects matching the bx is changing while objects are coming off
the iterator