Object sorters now use either a numeric or text sort explicitly based on the data type, instead of a generic sort.
This gets around a problem that perl >= 5.10 treats text strings
starting with "inf" as "infinity", including stringt like "info". :(

It is necessary in any case though, because this will take test which
has some numbers in it and mutually sort those as numbers.

  ($a <=> $b) || ($a cmp $b)

This is friendly for a user, but inconsistent with how a database will sort
a text column, and we require local and remote iterators to be in sync.