column_index_for_class_and_property renamed to column_index_for_class_property_and_object_num()
Takes a loading template's object number as an additional param.

The object num param is really required because a class/property pair isn't
always enough to uniquely identify a column within a particular object
fabricator.  For example, in the case where a single class is used for
multiple kinds of hangoff data of another class:

   class Thing {
       has => [
           hangoffs => { is => 'Hangoff', ... }
           color    => { via => 'hangoffs', to => 'value', where => [key => 'color']},
           age      => { via => 'hangoffs', to => 'value', where => [key => 'age']},
           rank     => { via => 'hangoffs', to => 'value', where => [key => 'rank']},
       ]
    }
    class Hangoff {
        has => [
            thing => { is => 'Thing', id_by => 'thing_id' },
            key   => { is => 'String' },
            value => { is => 'String' }
        ]
    }

During a get() like this:
    my @things = Thing->get(-hints => ['color','age','rank']);
There are several columns in the resultset for class => HangOff, property => value,
but only one for each object num.