Detect when a non-object is passed into an id_by accessor and throw a useful exception
Passing an object into an id_by accessor is supported as a way of changing
the id-holding property of the target object.  If the caller passes in a non-object
(like a number or string), then throw an exception that describes the problem, instead
of complaining about Can't locate object method "id" via package "whatever".