Re-fix implied property bug
The actual problem was that the metadata for id-by-properties and the id-ed-by
properties are linked.  For the metadata that can be shared (is_optional,
is_transient,is_constant,... meta_id_ref_shared_properties), if it is
specified on one but not the other, then copy it from the defined value.  If
neither specifies a value, put in the default for both
----------------------