Add is_connected property to datasources to flag whether it is connected
Processes used to be able to subscribe to a fake signal "connect" that only
worked on RDBMS data sources.

is_connected is now set to true by both RDBMS and File data sources when the
handle is opened, and set to false when it is closed.  Processes now need to
subscribe to changes on is_connected