Fix parsing of command-line values
A prior commit broke things when an optional param was being passed
a negative number as a value.

Getopt::Long documents that when using ':' in the spec, then if that param's
value starts with "-" or "--", it will be treated as a seperate option.  This
hack changes the regex it uses to look to options to treat negative numbers as
values.